+++ src/main/java/com/allancleiton/SCP/domain/repository/PalletRepository.java
@@ -5,10 +5,10 @@ import com.allancleiton.SCP.domain.entities.Pallet;
 import java.util.List;
 
 public interface PalletRepository {
-    Pallet findById(Long id);
+    Pallet findById(Integer id);
     void save(Pallet pallet);
     List<Pallet> findAll();
-    List<Pallet> findAllByCode(Long code);
-    Boolean deleteById(Long id);
+    List<Pallet> findAllByCode(Integer code);
+    Boolean deleteById(Integer id);
 
 }
